#699 Add ML Kit support
