+++ demo/app/app.css
@@ -82,8 +82,7 @@ button {
 }
 
 .barcodescanner {
-  width: 320;
-  height: 260;
+  height: 420;
   margin-top: 16;
 }
