+++ src/mlkit/facedetection/index.android.ts
@@ -1,0 +1,106 @@
+import { ImageSource } from "tns-core-modules/image-source";
+import { MLKitDetectFacesOptions, MLKitDetectFacesResult } from "./";
+import { MLKitOptions } from "../index";
+import { MLKitFaceDetection as MLKitFaceDetectionBase } from "./facedetection-common";
+
+declare const com: any;
+
+export class MLKitFaceDetection extends MLKitFaceDetectionBase {
+
+  protected createDetector(): any {
+    return getFaceDetector();
+  }
+
+  protected createSuccessListener(): any {
+    return new com.google.android.gms.tasks.OnSuccessListener({
+      onSuccess: (faces) => {
+
+        const imageSource = new ImageSource();
+        imageSource.setNativeSource(this.lastVisionImage.getBitmapForDebugging());
+
+        const result = <MLKitDetectFacesResult>{
+          imageSource: imageSource,
+          faces: []
+        };
+
+        // see https://github.com/firebase/quickstart-android/blob/0f4c86877fc5f771cac95797dffa8bd026dd9dc7/mlkit/app/src/main/java/com/google/firebase/samples/apps/mlkit/textrecognition/TextRecognitionProcessor.java#L62
+        for (let i = 0; i < faces.size(); i++) {
+          const face = faces.get(i);
+          result.faces.push({
+            smilingProbability: face.getSmilingProbability() !== com.google.firebase.ml.vision.face.FirebaseVisionFace.UNCOMPUTED_PROBABILITY ? face.getSmilingProbability() : undefined,
+            leftEyeOpenProbability: face.getLeftEyeOpenProbability() !== com.google.firebase.ml.vision.face.FirebaseVisionFace.UNCOMPUTED_PROBABILITY ? face.getLeftEyeOpenProbability() : undefined,
+            rightEyeOpenProbability: face.getRightEyeOpenProbability() !== com.google.firebase.ml.vision.face.FirebaseVisionFace.UNCOMPUTED_PROBABILITY ? face.getRightEyeOpenProbability() : undefined
+          });
+        }
+
+        this.notify({
+          eventName: MLKitFaceDetection.scanResultEvent,
+          object: this,
+          value: result
+        });
+      }
+    });
+  }
+}
+
+function getFaceDetector(): any {
+  const faceDetectorOptions =
+      new com.google.firebase.ml.vision.face.FirebaseVisionFaceDetectorOptions.Builder()
+          .setModeType(com.google.firebase.ml.vision.face.FirebaseVisionFaceDetectorOptions.ACCURATE_MODE)
+          .setLandmarkType(com.google.firebase.ml.vision.face.FirebaseVisionFaceDetectorOptions.ALL_LANDMARKS)
+          .setClassificationType(com.google.firebase.ml.vision.face.FirebaseVisionFaceDetectorOptions.ALL_CLASSIFICATIONS)
+          .setMinFaceSize(0.15)
+          .setTrackingEnabled(false)
+          .build();
+
+  return com.google.firebase.ml.vision.FirebaseVision.getInstance().getVisionFaceDetector(faceDetectorOptions);
+}
+
+export function detectFaces(options: MLKitDetectFacesOptions): Promise<MLKitDetectFacesResult> {
+  return new Promise((resolve, reject) => {
+    try {
+      const firebaseVisionFaceDetector = getFaceDetector();
+
+      const onSuccessListener = new com.google.android.gms.tasks.OnSuccessListener({
+        onSuccess: (faces) => {
+
+          const result = <MLKitDetectFacesResult>{
+            faces: []
+          };
+
+          // see https://github.com/firebase/quickstart-android/blob/0f4c86877fc5f771cac95797dffa8bd026dd9dc7/mlkit/app/src/main/java/com/google/firebase/samples/apps/mlkit/textrecognition/TextRecognitionProcessor.java#L62
+          for (let i = 0; i < faces.size(); i++) {
+            const face = faces.get(i);
+            result.faces.push({
+              smilingProbability: face.getSmilingProbability() !== com.google.firebase.ml.vision.face.FirebaseVisionFace.UNCOMPUTED_PROBABILITY ? face.getSmilingProbability() : undefined,
+              leftEyeOpenProbability: face.getLeftEyeOpenProbability() !== com.google.firebase.ml.vision.face.FirebaseVisionFace.UNCOMPUTED_PROBABILITY ? face.getLeftEyeOpenProbability() : undefined,
+              rightEyeOpenProbability: face.getRightEyeOpenProbability() !== com.google.firebase.ml.vision.face.FirebaseVisionFace.UNCOMPUTED_PROBABILITY ? face.getRightEyeOpenProbability() : undefined
+            });
+          }
+
+          resolve(result);
+          firebaseVisionFaceDetector.close();
+        }
+      });
+
+      const onFailureListener = new com.google.android.gms.tasks.OnFailureListener({
+        onFailure: exception => reject(exception.getMessage())
+      });
+
+      firebaseVisionFaceDetector
+          .detectInImage(getImage(options))
+          .addOnSuccessListener(onSuccessListener)
+          .addOnFailureListener(onFailureListener);
+
+    } catch (ex) {
+      console.log("Error in firebase.mlkit.detectFaces: " + ex);
+      reject(ex);
+    }
+  });
+}
+
+// TODO move up
+function getImage(options: MLKitOptions): any /* com.google.firebase.ml.vision.common.FirebaseVisionImage */ {
+  const image: android.graphics.Bitmap = options.image instanceof ImageSource ? options.image.android : options.image.imageSource.android;
+  return com.google.firebase.ml.vision.common.FirebaseVisionImage.fromBitmap(image);
+}
