+++ src/mlkit/index.ts
@@ -1,5 +1,6 @@
 import * as textrecognition from "./textrecognition";
 import * as barcodescanning from "./barcodescanning";
+import * as facedetection from "./facedetection";
 import { ImageSource } from "tns-core-modules/image-source";
 import { Image } from "tns-core-modules/ui/image";
 
@@ -11,8 +12,20 @@ export interface MLKitMultiEngineOptions extends MLKitOptions {
   engine?: "device" | "cloud"; // defaults to local, as cloud requires an additional plan
 }
 
+export interface MLKitResult {
+  /**
+   * Only set when using a camera stream,
+   * because when detecting from static images... you already have the image, right? :)
+   */
+  imageSource?: ImageSource;
+}
+
+export declare class MLKitBarcodeScanner {
+}
+
 export {
   textrecognition,
-  barcodescanning
+  barcodescanning,
+  facedetection
 };
